xilem_html: Move Either into its own module
